fix(#38): fixing issue with Google closure compiler and minified scripts
Clientlib-author was changed to be a wrapper clientlib and injecting FFY scripts manually.
We need to do it like that as Google Closure Compiler is having issues with already minified scripts
when adding those directly to cq.authoring.dialog clientlib. Options to disable gcc will not help as other clientlibs
of same category have minification turned on.

On top of that css.txt was deleted as it was doing nothing.